Make sure GD is installed, or fail
FossilOrigin-Name: 1a7ee46f325c60779f57162dc01505f4de3006b8

+++ dompdf/lib/PDF.php
@@ -24,6 +24,10 @@ class PDF
 
 	static public function install(): void
 	{
+		if (!function_exists('imagecreatefromwebp')) {
+			throw new \LogicException('You need to install the PHP GD extension to be able to use this extension.');
+		}
+
 		// Remove old setup, new one is shared between setups
 		if (file_exists(CACHE_ROOT . '/dompdf')) {
 			Utils::deleteRecursive(CACHE_ROOT . '/dompdf', true);
